Bank_Project/requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.15.4
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
virtualenv==20.26.3
